+++ tests/testthat/test_gho.R
@@ -48,3 +48,80 @@ test_that(
     expect_error(get_who_mr(age = 0:99, sex = rep(c("MLE", "FMLE"), 50), country = "XXXX"))
     expect_error(get_who_mr(age = 0:99, sex = rep(c("MLE", "FMLE"), 50), country = "FRA", year = 2050))
   })
+
+test_that(
+  "Conversion functions work", {
+    sex_char <- c("MLE", "FMLE", "FMLE", "MLE", "FMLE")
+    sex_01 <- c(0, 1, 1, 0, 1)
+    sex_12 <- c(1, 2, 2, 1, 2)
+    expect_identical(
+      heemod:::trans_sex_gho(sex_char),
+      sex_char
+    )
+    expect_identical(
+      heemod:::trans_sex_gho(sex_01),
+      sex_char
+    )
+    expect_identical(
+      heemod:::trans_sex_gho(sex_12),
+      sex_char
+    )
+    expect_error(
+      heemod:::trans_sex_gho(c(1, NA))
+    )
+    expect_error(
+      heemod:::trans_sex_gho(numeric())
+    )
+    expect_error(
+      heemod:::trans_sex_gho(1:3)
+    )
+    expect_error(
+      heemod:::trans_sex_gho(3:4)
+    )
+    expect_error(
+      heemod:::trans_sex_gho(c("xxx"))
+    )
+    expect_error(
+      heemod:::trans_sex_gho(NULL)
+    )
+    
+    age <- 0:101
+    age_res <- c(
+      "AGELT1", "AGE1-4", "AGE1-4", "AGE1-4", "AGE1-4", "AGE5-9", 
+      "AGE5-9", "AGE5-9", "AGE5-9", "AGE5-9", "AGE10-14", "AGE10-14", 
+      "AGE10-14", "AGE10-14", "AGE10-14", "AGE15-19", "AGE15-19", "AGE15-19", 
+      "AGE15-19", "AGE15-19", "AGE20-24", "AGE20-24", "AGE20-24", "AGE20-24", 
+      "AGE20-24", "AGE25-29", "AGE25-29", "AGE25-29", "AGE25-29", "AGE25-29", 
+      "AGE30-34", "AGE30-34", "AGE30-34", "AGE30-34", "AGE30-34", "AGE35-39", 
+      "AGE35-39", "AGE35-39", "AGE35-39", "AGE35-39", "AGE40-44", "AGE40-44", 
+      "AGE40-44", "AGE40-44", "AGE40-44", "AGE45-49", "AGE45-49", "AGE45-49", 
+      "AGE45-49", "AGE45-49", "AGE50-54", "AGE50-54", "AGE50-54", "AGE50-54", 
+      "AGE50-54", "AGE55-59", "AGE55-59", "AGE55-59", "AGE55-59", "AGE55-59", 
+      "AGE60-64", "AGE60-64", "AGE60-64", "AGE60-64", "AGE60-64", "AGE65-69", 
+      "AGE65-69", "AGE65-69", "AGE65-69", "AGE65-69", "AGE70-74", "AGE70-74", 
+      "AGE70-74", "AGE70-74", "AGE70-74", "AGE75-79", "AGE75-79", "AGE75-79", 
+      "AGE75-79", "AGE75-79", "AGE80-84", "AGE80-84", "AGE80-84", "AGE80-84", 
+      "AGE80-84", "AGE85-89", "AGE85-89", "AGE85-89", "AGE85-89", "AGE85-89", 
+      "AGE90-94", "AGE90-94", "AGE90-94", "AGE90-94", "AGE90-94", "AGE95-99", 
+      "AGE95-99", "AGE95-99", "AGE95-99", "AGE95-99", "AGE100PLUS", 
+      "AGE100PLUS"
+    )
+    
+    expect_identical(
+      heemod:::trans_age_gho(age),
+      age_res
+    )
+    expect_error(
+      heemod:::trans_age_gho(-1)
+    )
+    expect_error(
+      heemod:::trans_age_gho(c(NA, 1))
+    )
+    expect_error(
+      heemod:::trans_age_gho(numeric())
+    )
+    expect_error(
+      heemod:::trans_age_gho(NULL)
+    )
+  }
+)